+++ R/replace_regions_with_countries.R
@@ -9,7 +9,7 @@ replace_regions_with_countries <- function(gdp, unit_in, base_x, with_regions, s
   gdp <- dplyr::filter(gdp, ! .data$iso3c %in% my_reg)
 
   # Disaggregate regions
-  weight_unit <- stringr::str_match(unit_in, "\\$(...)")[2]
+  weight_unit <- sub("\\$", "", regmatches(unit_in, regexpr("\\$(...)", unit_in)))
   weight_year <- base_x
   with_regions <- dplyr::rename(with_regions, "gdpuc_region" = .data$region)
   gdp_reg <- disaggregate_regions(gdp_reg, with_regions, weight_unit, weight_year, source)
@@ -28,7 +28,7 @@ disaggregate_regions <- function(gdp, with_regions, weight_unit, weight_year, so
   regex_var <- "GDP, PPP \\(constant .... international \\$\\)"
   regex_year <- "GDP, PPP \\(constant (....) international \\$\\)"
   share_var <- grep(regex_var, colnames(source), value = TRUE)[1]
-  share_year <- stringr::str_match(share_var, regex_year)[, 2]
+  share_year <- regmatches(share_var, regexpr("[[:digit:]]{4}", share_var))
   unit_in <- paste("constant", share_year, "Int$PPP")
 
   # Convert that variable to desired unit, and compute shares of GDP per region